+++ packages/sdk-metrics/test/performance/benchmarks/otel-sdk-metrics-paths.cjs
@@ -1,20 +1,36 @@
+/*
+ * Copyright The OpenTelemetry Authors
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
 const os = require('node:os');
 const { performance } = require('node:perf_hooks');
 const { ROOT_CONTEXT } = require('@opentelemetry/api');
 
-const { AsyncMetricStorage } = require('../../packages/sdk-metrics/build/src/state/AsyncMetricStorage.js');
-const { MultiMetricStorage } = require('../../packages/sdk-metrics/build/src/state/MultiWritableMetricStorage.js');
-const { AttributeHashMap } = require('../../packages/sdk-metrics/build/src/state/HashMap.js');
+const { AsyncMetricStorage } = require('../../../build/src/state/AsyncMetricStorage.js');
+const { MultiMetricStorage } = require('../../../build/src/state/MultiWritableMetricStorage.js');
+const { AttributeHashMap } = require('../../../build/src/state/HashMap.js');
 const {
   createAllowListAttributesProcessor,
-} = require('../../packages/sdk-metrics/build/src/view/AttributesProcessor.js');
-const { SumAggregator } = require('../../packages/sdk-metrics/build/src/aggregator/Sum.js');
+} = require('../../../build/src/view/AttributesProcessor.js');
+const { SumAggregator } = require('../../../build/src/aggregator/Sum.js');
 const {
   createInstrumentDescriptor,
-} = require('../../packages/sdk-metrics/build/src/InstrumentDescriptor.js');
+} = require('../../../build/src/InstrumentDescriptor.js');
 const {
   InstrumentType,
-} = require('../../packages/sdk-metrics/build/src/export/MetricData.js');
+} = require('../../../build/src/export/MetricData.js');
 
 let sink = 0;
 
@@ -74,7 +90,9 @@ const attributeProcessor = createAllowListAttributesProcessor([
   'region',
 ]);
 
-const processorInput = Array.from({ length: 4_000 }, (_, i) => createAttributes(i));
+const processorInput = Array.from({ length: 4_000 }, (_, i) =>
+  createAttributes(i)
+);
 
 const allowListProcessorBench = benchmark(
   'AllowListProcessor.process x 4k',